requirements.txt
asgiref==3.5.2
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
sqlparse==0.4.2
psycopg2-binary==2.9.3
pytest-django==4.5.2